moving new scripts processing_mgmt and run_tools to their own dir
